Remove pack from drafter after a pick

+++ src/main/kotlin/com/tobert/cube/repositories/DraftRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface DraftRepository : JpaRepository<Draft, Long>
+interface DraftRepository : JpaRepository<Draft, Int>
